+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * BraintreeDataCollector
   * Fix intermittent crash in `collectDeviceData` caused by an internal SDK component accessing `UIPasteboard` off the main thread
+* BraintreePayPal
+  * Fix `BTPayPalAccountNonce.clientMetadataID` returning `nil` caused by `clientMetadataID` being generated off the main thread
 
 ## 7.6.0 (2026-05-18)
 * Fix inconsistency in minimum deployment target, which is now consistently iOS 16 (fixes #1757)