+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   * Fix intermittent crash in `collectDeviceData` caused by an internal SDK component accessing `UIPasteboard` off the main thread
 * BraintreePayPal
   * Fix `BTPayPalAccountNonce.clientMetadataID` returning `nil` caused by `clientMetadataID` being generated off the main thread
+  * Fix `BTPayPalAccountNonce.clientMetadataID` returning `nil` for checkout flows with `requestBillingAgreement` enabled
 
 ## 7.6.0 (2026-05-18)
 * Fix inconsistency in minimum deployment target, which is now consistently iOS 16 (fixes #1757)